Fixes
Upon investigating whether the code could be written in a loop, the
issue of dynamically creating variable names for both tables and for
running the dynatable searches came up, and I wasn't able to find a
workaround, so the status.jade file remains as 2 concrete tables.
There's no efficiency lost however, since there will only ever be two
tables in that page.  I did remove the unnecessary scripts in the block
scripts section of the file however.

Also modified layout.jade to include the status page in the navbar of
the site under the name "Accepted Queue"